[FIX] convene: Delete user wasn't working if deleting more than one user due to page not reloading (so reload), also ask user before removing a row
git-svn-id: https://svn.code.sf.net/p/tikiwiki/code/branches/12.x@48074 b456876b-0849-0410-b77d-98878d47e9d5

+++ lib/wiki-plugins/wikiplugin_convene.php
@@ -297,7 +297,7 @@ FORM;
 
 				this.data = data;
 
-				this.save();
+				this.save(true);
 			},
 			addDate: function(date) {
 				if (!date) return;
@@ -339,9 +339,10 @@ FORM;
 				this.data = newData.join('$n');
 				this.save();
 			},
-			save: function() {
+			save: function(reload) {
 				$("#page-data").modal(tr("Loading..."));
 
+				var needReload = reload != undefined;
 				var params = {
 					page: "$page",
 					content: $.trim(this.data),
@@ -355,12 +356,16 @@ FORM;
 				};
 				$.post("tiki-wikiplugin_edit.php", params, function() {
 					$.get($.service("wiki", "get_page", {page: "$page"}), function (data) {
-						if (data) {
-							var newForm = $("#pluginConvene$i", data);
-							$("#pluginConvene$i", "#page-data").replaceWith(newForm);
+						if (needReload) {
+							history.go(0);
+						} else {
+							if (data) {
+								var newForm = $("#pluginConvene$i", data);
+								$("#pluginConvene$i", "#page-data").replaceWith(newForm);
+							}
+							initConvene$i();
+							$("#page-data").modal();
 						}
-						initConvene$i();
-						$("#page-data").modal();
 					});
 
 				});
@@ -431,7 +436,10 @@ FORM;
 
 			$('.conveneDeleteUser$i')
 				.click(function() {
-					convene$i.deleteUser($(this).data("user"));
+					if (confirm(tr("Are you sure you want to remove this user's votes?") + "\\n" +
+							tr("There is no undo"))) {
+						convene$i.deleteUser($(this).data("user"));
+					}
 					return false;
 				});
